+++ src/IOperand/types/int16.cpp
@@ -59,8 +59,8 @@ avm::IOperand *avm::myInt16::operator+(const IOperand &other) const
     avm::Factory fct;
     avm::myException exc;
 
-    int16_t thisValue = static_cast<int16_t>(std::stod(this->value));
-    long double otherValue = static_cast<long double>(std::stod(other.toString()));
+    int16_t thisValue = static_cast<int16_t>(std::stold(this->value));
+    long double otherValue = static_cast<long double>(std::stold(other.toString()));
     if (this->type >= other.getType()) {
         if (isAddOverflowing(std::numeric_limits<int16_t>::max(), std::numeric_limits<int16_t>::min(), thisValue, otherValue)) {
             exc.printError("myInt16::operator+ error: overflow or underflow\n");
@@ -87,8 +87,8 @@ avm::IOperand *avm::myInt16::operator-(const IOperand &other) const
     avm::Factory fct;
     avm::myException exc;
 
-    int16_t thisValue = static_cast<int16_t>(std::stod(this->value));
-    long double otherValue = static_cast<long double>(std::stod(other.toString()));
+    int16_t thisValue = static_cast<int16_t>(std::stold(this->value));
+    long double otherValue = static_cast<long double>(std::stold(other.toString()));
     if (this->type >= other.getType()) {
         if (isSubOverflowing(std::numeric_limits<int16_t>::max(), std::numeric_limits<int16_t>::min(), thisValue, otherValue)) {
             exc.printError("myInt16::operator+ error: overflow or underflow\n");
@@ -115,8 +115,8 @@ avm::IOperand *avm::myInt16::operator*(const IOperand &other) const
     avm::Factory fct;
     avm::myException exc;
 
-    int16_t thisValue = static_cast<int16_t>(std::stod(this->value));
-    long double otherValue = static_cast<long double>(std::stod(other.toString()));
+    int16_t thisValue = static_cast<int16_t>(std::stold(this->value));
+    long double otherValue = static_cast<long double>(std::stold(other.toString()));
     if (this->type >= other.getType()) {
         if (isMulOverflowing(std::numeric_limits<int16_t>::max(), std::numeric_limits<int16_t>::min(), thisValue, otherValue)) {
             exc.printError("myInt16::operator+ error: overflow or underflow\n");
@@ -143,8 +143,8 @@ avm::IOperand *avm::myInt16::operator/(const IOperand &other) const
     avm::Factory fct;
     avm::myException exc;
 
-    int16_t thisValue = static_cast<int16_t>(std::stod(this->value));
-    int64_t otherValue = static_cast<int64_t>(std::stod(other.toString()));
+    int16_t thisValue = static_cast<int16_t>(std::stold(this->value));
+    int64_t otherValue = static_cast<int64_t>(std::stold(other.toString()));
     if (thisValue == 0) {
         exc.printError("myInt16::operator/ error: division by 0\n");
         return (NULL);
@@ -164,8 +164,8 @@ avm::IOperand *avm::myInt16::operator%(const IOperand &other) const
     avm::Factory fct;
     avm::myException exc;
 
-    int16_t thisValue = static_cast<int16_t>(std::stod(this->value));
-    int64_t otherValue = static_cast<int64_t>(std::stod(other.toString()));
+    int16_t thisValue = static_cast<int16_t>(std::stold(this->value));
+    int64_t otherValue = static_cast<int64_t>(std::stold(other.toString()));
     if (thisValue == 0) {
         exc.printError("myInt16::operator\% error: modulo by 0\n");
         return (NULL);
