+++ src/IOperand/types/int32.cpp
@@ -58,8 +58,8 @@ avm::IOperand *avm::myInt32::operator+(const IOperand &other) const
     avm::Factory fct;
     avm::myException exc;
 
-    int32_t thisValue = static_cast<int32_t>(std::stod(this->value));
-    long double otherValue = static_cast<long double>(std::stod(other.toString()));
+    int32_t thisValue = static_cast<int32_t>(std::stold(this->value));
+    long double otherValue = static_cast<long double>(std::stold(other.toString()));
     if (this->type >= other.getType()) {
         if (isAddOverflowing(std::numeric_limits<int32_t>::max(), std::numeric_limits<int32_t>::min(), thisValue, otherValue)) {
             exc.printError("myInt32::operator+ error: overflow or underflow\n");
@@ -86,8 +86,8 @@ avm::IOperand *avm::myInt32::operator-(const IOperand &other) const
     avm::Factory fct;
     avm::myException exc;
 
-    int32_t thisValue = static_cast<int32_t>(std::stod(this->value));
-    long double otherValue = static_cast<long double>(std::stod(other.toString()));
+    int32_t thisValue = static_cast<int32_t>(std::stold(this->value));
+    long double otherValue = static_cast<long double>(std::stold(other.toString()));
     if (this->type >= other.getType()) {
         if (isSubOverflowing(std::numeric_limits<int32_t>::max(), std::numeric_limits<int32_t>::min(), thisValue, otherValue)) {
             exc.printError("myInt32::operator+ error: overflow or underflow\n");
@@ -114,8 +114,8 @@ avm::IOperand *avm::myInt32::operator*(const IOperand &other) const
     avm::Factory fct;
     avm::myException exc;
 
-    int32_t thisValue = static_cast<int32_t>(std::stod(this->value));
-    long double otherValue = static_cast<long double>(std::stod(other.toString()));
+    int32_t thisValue = static_cast<int32_t>(std::stold(this->value));
+    long double otherValue = static_cast<long double>(std::stold(other.toString()));
     if (this->type >= other.getType()) {
         if (isMulOverflowing(std::numeric_limits<int32_t>::max(), std::numeric_limits<int32_t>::min(), thisValue, otherValue)) {
             exc.printError("myInt32::operator+ error: overflow or underflow\n");
@@ -143,8 +143,8 @@ avm::IOperand *avm::myInt32::operator/(const IOperand &other) const
     avm::myException exc;
 
 
-    int32_t thisValue = static_cast<int32_t>(std::stod(this->value));
-    int64_t otherValue = static_cast<int64_t>(std::stod(other.toString()));
+    int32_t thisValue = static_cast<int32_t>(std::stold(this->value));
+    int64_t otherValue = static_cast<int64_t>(std::stold(other.toString()));
     if (thisValue == 0) {
         exc.printError("myint32::operator/ error: division by 0\n");
         return (NULL);
@@ -164,8 +164,8 @@ avm::IOperand *avm::myInt32::operator%(const IOperand &other) const
     avm::Factory fct;
     avm::myException exc;
 
-    int32_t thisValue = static_cast<int32_t>(std::stod(this->value));
-    int64_t otherValue = static_cast<int64_t>(std::stod(other.toString()));
+    int32_t thisValue = static_cast<int32_t>(std::stold(this->value));
+    int64_t otherValue = static_cast<int64_t>(std::stold(other.toString()));
     if (thisValue == 0) {
         exc.printError("myint32::operator\% error: modulo by 0\n");
         return (NULL);
